worked on city

+++ src/app/SystemAdministration/country/country.component.ts
@@ -84,11 +84,12 @@ export class CountryComponent implements OnInit {
   selectedRowCity: any[];
   nodeCitySelect: string;
   saveUpdateCity: string;
+  saveUpdateAddressPopup: string;
+  btnSaveUpdateAddressPopup: string;
 
   selectedRowPostal: any[];
   nodePostalSelect: string;
   saveUpdatePostal: string;
-
 
   TotalCountry: any = 0;
   StartCountryNumber: any = 0;
@@ -99,6 +100,8 @@ export class CountryComponent implements OnInit {
 
   TotalCity: any = 0;
   ShowLimitedCity: any = 0;
+  ShowCity: any = 0;
+  selectAllCity = true;
 
   TotalPostal: any = 0;
   ShowLimitedPostal: any = 0;
@@ -389,7 +392,7 @@ export class CountryComponent implements OnInit {
     this.deleteStateToggleButton = true;
 
     this.saveUpdateCity = 'Save';
-    this.addCityToggleButton = false;
+    this.addCityToggleButton = true;
     this.saveCityToggleButton = true;
     this.deleteCityToggleButton = true;
 
@@ -874,30 +877,38 @@ export class CountryComponent implements OnInit {
 
   getCity() {
     const universalBody = new UniversalBody();
-    universalBody.userID = '1'; this.allWeb.getCity(universalBody)
+    universalBody.userID = '1';
+   
+    this.isShowing = true;
+    this.allWeb.getCity(universalBody)
       .subscribe(
         data => {
+          this.isShowing = false;
           this.getCityResponse = data;
 
           if (this.getCityResponse.length === 0) {
 
-            this.TotalCity = this.getCityResponse.length;
+            //this.TotalCity = this.getCityResponse.length;
             this.saveUpdateCity = "Save";
 
             this.addCityToggleButton = false;
             this.saveCityToggleButton = true;
             this.deleteCityToggleButton = true;
+            this.selectAllCity = true;
           } else {
+            this.isShowing = false;
 
             if (this.getCityResponse.length >= 50) {
               this.ShowLimitedCity = 50;
+              this.ShowCity = 1;
             } else {
               this.ShowLimitedCity = this.getCityResponse.length;
+              this.ShowCity = 1;
             }
 
             this.TotalCity = this.getCityResponse.length;
             this.saveUpdateCity = "Save";
-
+            this.selectAllCity = false;
             this.saveCityToggleButton = true;
             this.addCityToggleButton = false;
             this.deleteCityToggleButton = true;
@@ -1064,6 +1075,42 @@ export class CountryComponent implements OnInit {
 
   }
 
+  onAddState() {
+    this.saveStateToggleButton = false;
+    this.nodeStateSelect = "Add";
+    //  this.addStateToggleButton = true;
+    this.saveUpdateState = 'Save';
+    this.stateApi.setFocusedCell(this.countState, 'countryName');
+    this.stateApi.getColumnDef('stateName').editable = true;
+    this.stateApi.getColumnDef('description').editable = true;
+    this.countState++;
+    let res = this.stateApi.updateRowData({ add: [{ stateName: '', countryName: '', description: '', hidden: 'hidden' }], addIndex: 0 });
+    const universalJsonBody = new UniversalJsonBody();
+    const selectedNodes = this.stateApi.getSelectedNodes();
+    const selectedData = selectedNodes.map(node => node.data);
+    var dataTest: Object;
+    selectedData.map(node => dataTest = node as Object);
+
+    if (dataTest['stateName'] === '') {
+      alert('Enter state name');
+    }
+    else {
+      for (let selectedNode of selectedData) {
+        const stateBody = new StateBody();
+        stateBody.stateName = selectedNode['stateName'];
+        stateBody.description = selectedNode['description'];
+        stateBody.countryId = this.selectedCountryId + '';
+        this.stateArray.push(stateBody);
+
+      }
+
+    }
+
+
+
+  }
+
+
   onAddCity() {
     this.saveCityToggleButton = false;
     this.nodeCitySelect = "Add";
@@ -1081,7 +1128,7 @@ export class CountryComponent implements OnInit {
     const selectedData = selectedNodes.map(node => node.data);
     var dataTest: Object;
     selectedData.map(node => dataTest = node as Object);
-
+    this.cityApi.tabToNextCell();
     if (dataTest['cityName'] === '') {
       alert('Enter city name');
     }
@@ -1093,13 +1140,9 @@ export class CountryComponent implements OnInit {
         cityBody.stateId = this.selectedStateId + '';
         this.cityArray.push(cityBody);
         var jsonData = JSON.stringify(this.cityArray);
-
+        jsonData = jsonData.replace(/"/g, "'");
       }
-
-
     }
-
-
   }
 
   onAddPostal() {
@@ -1147,6 +1190,23 @@ export class CountryComponent implements OnInit {
   }
 
   universalSaveAddress() {
+    // const countrySelectedNodes = this.api.getSelectedNodes();
+    // const stateSelectedNodes = this.stateApi.getSelectedNodes();
+    const citySelectedNodes = this.cityApi.getSelectedNodes();
+    // const postalSelectedNodes = this.postalApi.getSelectedNodes();
+    // if (countrySelectedNodes.length !== 0) {
+    //   this.onDeleteCountry();
+    // }
+    // else if (stateSelectedNodes.length !== 0) {
+    //   this.onDeleteState();
+    // }
+    if (citySelectedNodes.length !== 0) {
+      this.onSaveUpdateCity();
+    }
+    // else if (postalSelectedNodes.length !== 0) {
+    //   this.onDeletePostal();
+    // }
+
      const countrySelectedNodes = this.api.getSelectedNodes();
     // const stateSelectedNodes = this.stateApi.getSelectedNodes();
     const citySelectedNodes = this.cityApi.getSelectedNodes();
@@ -1244,11 +1304,13 @@ export class CountryComponent implements OnInit {
   onDeleteCity() {
 
     const selectedNodes = this.cityApi.getSelectedNodes();
-    const deleteCityArray: DeleteCityBody[] = [];
-
+    var deleteCityArray: DeleteCityBody[] = [];
+   // var universlResponse = new UniversalResponse();
     const universalJsonBody = new UniversalJsonBody();
     const selectedData = selectedNodes.map(node => node.data);
     var dataTest: Object;
+    var output:string = "";
+    var count=0;
 
     selectedData.map(node => dataTest = node as Object);
     if (selectedData.length === 0) {
@@ -1259,8 +1321,16 @@ export class CountryComponent implements OnInit {
       for (let selectedNode of selectedData) {
         const deleteCityBody = new DeleteCityBody();
         deleteCityBody.cityId = selectedNode['cityID'];
+        if (deleteCityBody.cityId === undefined) {
+          this.cityApi.updateRowData({ remove: selectedData });
+        } else {
+          deleteCityBody.userId = '1';
+          deleteCityArray.push(deleteCityBody);
+          var jsonData = JSON.stringify(deleteCityArray);
+        }
 
-        deleteCityArray.push(deleteCityBody);
+        //  // if(deleteCityBody.cityId === 'undefine')
+        //   deleteCityArray.push(deleteCityBody);
 
       }
 
@@ -1268,24 +1338,61 @@ export class CountryComponent implements OnInit {
       jsonData1 = jsonData1.replace(/"/g, "'");
 
       universalJsonBody.jsonData = jsonData1;
-      this.allWeb.deleteCity(universalJsonBody)
-        .subscribe(
-          data => {
-            this.universalResponse = data;
-            alert(this.universalResponse.MESSAGE);
-            if (this.universalResponse.STATUS === 'Success') {
+      if (deleteCityArray.length === 0) {
+        this.cityFilter = false;
+      }
+      else {
+        this.allWeb.deleteCity(universalJsonBody)
+          .subscribe(
+            data => {
+              // universlResponse = data;
+              
+              // if (universlResponse.STATUS.trim() === 'Success') {
+              //   if(universlResponse.OUTPUT != '') {
+              //     var names:string[] = universlResponse.OUTPUT.split(",");         
+              //     for(var i = 0;i<names.length-1;i++) { 
+              //       output += names[i]+"\n";
+              //     }
+              //      alert(output);   
+              //   }else{
+              //     alert(universlResponse.MESSAGE);
+              //     this.getCity();
+              //     this.cityApi.removeItems(selectedNodes);
+              //   }
+              // }
+            
+              this.universalResponse= data;
+              if (this.universalResponse.STATUS === 'Success') {
+                if(this.universalResponse.OUTPUT != '') {
+                  var names:string[] = this.universalResponse.OUTPUT.split(",");         
+                  for(var i = 0;i<names.length-1;i++) { 
+                    output += names[i]+"\n";
+                    count++;
+                  }
+                   alert(output);   
+                }else if(deleteCityArray.length === count){
+                  alert(this.universalResponse.MESSAGE);
+                  this.getCity();
+                 this.cityApi.removeItems(selectedNodes);
+                }
+              }
+              // alert(universlResponse.MESSAGE);
+              // if (universlResponse.STATUS === 'Success') {
 
-              this.cityApi.removeItems(selectedNodes);
-
-              this.getCity();
-            }
-          }
-        );
-
+              //   this.cityApi.removeItems(selectedNodes);
+              //   deleteCityArray = [];
+              //   this.getCity();
+              //   this.cityFilter = false;
+              // }
+             }
+          );
+      }
     }
-    this.addCityToggleButton = false;
-    this.saveCityToggleButton = true;
-    this.deleteCityToggleButton = true;
+
+   // this.cityApi.removeItems(selectedNodes);
+    // this.addCityToggleButton = false;
+    // this.saveCityToggleButton = true;
+    // this.deleteCityToggleButton = true;
 
   }
 
@@ -1396,6 +1503,7 @@ export class CountryComponent implements OnInit {
 
     }
   }
+
 
   onCityFilterChange() {
 
@@ -1599,11 +1707,13 @@ export class CountryComponent implements OnInit {
   // }
 
   public selectedStateId: number;
-  public selectedCityId: number
+  public selectedCityId: number;
+
   onSaveCity() {
     const universalJsonBody = new UniversalJsonBody();
 
     const selectedNodes = this.cityApi.getSelectedNodes();
+    var output:string ='';
 
     const selectedData = selectedNodes.map(node => node.data);
     var dataTest: Object;
@@ -1627,25 +1737,61 @@ export class CountryComponent implements OnInit {
           console.log('jsondata:   ' + jsonData);
         }
         universalJsonBody.jsonData = jsonData;
+        this.isShowing = true;
         this.allWeb.saveCity(universalJsonBody)
           .subscribe(
 
             data => {
+              this.isShowing = false;
               this.universalResponse = data;
 
-              alert(this.universalResponse.MESSAGE);
-
               if (this.universalResponse.STATUS.trim() === 'Success') {
-
-                this.getCity();
-
+                if(this.universalResponse.OUTPUT != '') {
+                  var names:string[] = this.universalResponse.OUTPUT.split(",");         
+                  for(var i = 0;i<names.length-1;i++) { 
+                    output += names[i]+"\n";
+                  }
+                   alert(output);   
+                }else{
+                  alert(this.universalResponse.MESSAGE);
+                  this.getCity();
+                 
+                }
               }
-              this.cityArray = [];
+
+
+
+              // if (this.universalResponse.STATUS.trim() === 'Success') {
+              //   if (this.cityArray.length >= 1) {
+              //     if (this.universalResponse.OUTPUT === '') {
+              //       alert(this.universalResponse.MESSAGE);
+              //       this.getCity();
+              //     }
+              //     else {
+              //       alert(this.universalResponse.MESSAGE + ' ' + this.universalResponse.OUTPUT + ' ' + 'already existed ');
+              //     }
+              //   }
+              //   else {
+              //     if (this.universalResponse.OUTPUT === '') {
+              //       alert(this.universalResponse.MESSAGE);
+              //       this.getCity();
+              //     }
+              //     else {
+              //       alert(this.universalResponse.MESSAGE + ' ' + this.universalResponse.OUTPUT + ' ' + 'already existed');
+              //     }
+              //   }
+
+
+              //}
+
+
             }
           );
       }
     }
-
+    this.cityArray = [];
+  //  alert(this.cityArray.length);
+    this.nodeCitySelect = 'Update'
   }
 
   onCitySelectionChanged() {
@@ -1663,120 +1809,138 @@ export class CountryComponent implements OnInit {
         this.deleteCityToggleButton = false;
         this.cityFilter = true;
         this.cityCheckedStatus = false;
+        this.saveCityToggleButton = false;
       }
       else {
         this.deleteCityToggleButton = false;
         this.cityFilter = false;
         this.cityCheckedStatus = false;
-
+        this.saveCityToggleButton = false;
       }
       if (this.nodeCitySelect === 'Add') {
         this.saveUpdateCity = 'Save';
-        this.nodeCitySelect = 'Edit';
+        // this.nodeCitySelect = 'Edit';
+        this.saveUpdateAddressPopup = 'Do you want to save?';
+        this.btnSaveUpdateAddressPopup = 'SAVE';
 
       } else if (this.nodeCitySelect === undefined) {
         this.saveUpdateCity = 'Save';
         this.saveCityToggleButton = false;
+        this.saveUpdateAddressPopup = 'Do you want to update?';
+        this.btnSaveUpdateAddressPopup = 'UPDATE';
+      }
+      else if (this.nodeCitySelect === 'Update') {
+        this.saveCityToggleButton = false;
+        this.saveUpdateCity = 'Save';
+        this.btnSaveUpdateAddressPopup = 'UPDATE';
+        this.saveUpdateAddressPopup = 'Do you want to update?';
       }
     }
     else if (this.selectedRowCity.length >= 1) {
-      this.saveCityToggleButton = true;
-
+      this.saveCityToggleButton = false;
+      this.deleteCityToggleButton = false;
+      this.addCityToggleButton = false;
     }
-
-
-
-
-
-    // if (this.selectedRowCity.length === 1) {
-    //   if (this.getCityResponse.length === 1) {
-    //     this.deleteCityToggleButton = false;
-    //     this.cityFilter = true;
-    //     this.cityCheckedStatus = false;
-    //   }
-    //   else {
-    //     this.deleteCityToggleButton = false;
-    //     this.cityFilter = false;
-    //     this.cityCheckedStatus = false;
-    //   }
-    //   if (this.nodeCitySelect === "Add") {
-    //     this.saveUpdateCity = "Save";
-    //     this.nodeCitySelect = "Update";
-    //   } else if (this.nodeCitySelect === undefined) {
-    //     this.saveUpdateCity = "Update";
-
-    //     this.saveCityToggleButton = false;
-    //   }
-    //   else if (this.nodeCitySelect === 'Update') {
-    //     this.saveCityToggleButton = false;
-    //     this.saveUpdateCity = 'Update';
-
-    //   }
-    //   else if (this.selectedRowCity.length >= 1) {
-    //     this.saveCityToggleButton = true;
-
-    //   }
-    // }
   }
 
   onSaveUpdateCity() {
     this.cityApi.tabToNextCell();
-    if (this.saveUpdateCity === "Save") {
+    if (this.nodeCitySelect === "Add") {
       this.onSaveCity();
     }
-    else if (this.saveUpdateCity === "Update") {
+    else {
+
       this.onUpdateCity();
     }
-    this.addPostalToggleButton = false;
+    // this.addPostalToggleButton = false;
   }
 
   onUpdateCity() {
-    const updateCityBody = new UpdateCityBody();
+    this.cityArray = [];
+    const universalJsonBody = new UniversalJsonBody();
 
     const selectedNodes = this.cityApi.getSelectedNodes();
 
     const selectedData = selectedNodes.map(node => node.data);
     var dataTest: Object;
+    var output: string ='';
     selectedData.map(node => dataTest = node as Object);
     if (selectedData.length === 0) {
-      alert("Please select a row");
-    }
-    updateCityBody.CityName = dataTest['cityName'];
-    updateCityBody.Description = dataTest['description'];
-
-    if (dataTest['cityName'] === '') {
-      alert("Enter city name");
-
+      alert('Please select a row');
     }
     else {
-      updateCityBody.StateID = this.selectedStateId
-      updateCityBody.CityID = dataTest['cityID'];
-      if (updateCityBody.CityID === undefined) {
-
-        this.addCityToggleButton = false;
+      if (dataTest['cityName'] === '') {
+        alert('Enter city name');
       }
       else {
-
-        this.addCityToggleButton = false;
-
-        this.allWeb.updateCity(updateCityBody)
+        for (let selectedNode of selectedData) {
+          const cityBody = new CityBody();
+          cityBody.cityName = selectedNode['cityName'];
+          cityBody.description = selectedNode['description'];
+          cityBody.stateId = this.selectedStateId + '';
+          // updateCityBody.StateID = this.selectedStateId
+          cityBody.cityId =  selectedNode['cityID'];
+         // cityBody.cityId = selectedNode['cityID'];
+          this.cityArray.push(cityBody);
+          var jsonData = JSON.stringify(this.cityArray);
+          jsonData = jsonData.replace(/"/g, "'");
+          console.log('jsondata:   ' + jsonData);
+        }
+        universalJsonBody.jsonData = jsonData;
+        this.isShowing = true;
+        this.allWeb.updateCity(universalJsonBody)
           .subscribe(
+
             data => {
+              this.isShowing = false;
               this.universalResponse = data;
 
-              alert(this.universalResponse.MESSAGE);
-
-              if (this.universalResponse.STATUS === 'Success') {
-
-                this.getCity();
-                this.addCityToggleButton = false;
-
+              if (this.universalResponse.STATUS.trim() === 'Success') {
+                if(this.universalResponse.OUTPUT != '') {
+                  var names:string[] = this.universalResponse.OUTPUT.split(",");         
+                  for(var i = 0;i<names.length-1;i++) { 
+                    output += names[i]+"\n";
+                  }
+                   alert(output);   
+                }else{
+                  alert(this.universalResponse.MESSAGE);
+                  this.getCity();
+                 
+                }
               }
-            }
 
+              // if (this.universalResponse.STATUS.trim() === 'Success') {
+              //   if (this.cityArray.length >= 1) {
+              //     if (this.universalResponse.OUTPUT === '') {
+              //       alert(this.universalResponse.MESSAGE);
+              //       this.getCity();
+              //     }
+              //     else {
+              //       alert(this.universalResponse.MESSAGE + ' ' + this.universalResponse.OUTPUT + ' ' + 'already existed ');
+              //     }
+              //   }
+              //   else {
+              //     if (this.universalResponse.OUTPUT === '') {
+              //       alert(this.universalResponse.MESSAGE);
+              //       this.getCity();
+              //     }
+              //     else {
+              //       alert(this.universalResponse.MESSAGE + ' ' + this.universalResponse.OUTPUT + ' ' + 'already existed');
+              //     }
+              //   }
+
+
+              // }
+
+
+            }
           );
       }
     }
+    this.cityArray = [];
+   // alert(this.cityArray.length);
+    this.nodeCitySelect = 'Update'
+
   }
 
   onGridReady(params) {
@@ -1805,11 +1969,15 @@ export class CountryComponent implements OnInit {
         this.countryCheckedStatus = false;
         this.addCountryToggleButton = false;
       }
+
       else {
         this.deleteCountryToggleButton = false;
         this.countryFilter = false;
         this.countryCheckedStatus = false;
         this.addCountryToggleButton = false;
+      }
+      if (this.cityApi.getDisplayedRowCount() >= 1) {
+        this.countryFilter = false;
       }
       if (this.nodeCountrySelect === "Add") {
         this.saveUpdateCountry = "Save";
